deploy/requirements-deploy.txt
# Production deployment dependencies.
# Pin to match training environment (see PRODUCTION_DEPLOYMENT_GUIDE).
# timezonefinder + pgeocode for zip-based timezone lookup (optional; country-only works without).
# python-dotenv: optional for local dev (inference_job loads deploy/.env when present).

numpy>=1.24,<2
pandas>=2.0
# Match training env: model was pickled with sklearn 1.8.0 (InisotonicCalibratedClassifier).
scikit-learn>=1.8,<1.9
catboost>=1.2,<1.3
joblib>=1.3
google-cloud-bigquery>=3.0
pandas-gbq>=0.26.1
google-cloud-storage>=2.0
timezonefinder>=6.0
pgeocode>=0.4
python-dotenv>=1.0
